requirements.txt
opencv-python>=4.8.0
numpy>=1.20.0
Pillow>=10.0.0
cryptography>=41.0.0
click>=8.0.0